+++ src/main/java/com/candido/server/domain/v1/token/TokenRepository.java
@@ -9,12 +9,12 @@ import java.util.Optional;
 
 public interface TokenRepository extends JpaRepository<Token, Integer> {
     @EntityGraph(attributePaths = {"tokenScopeCategory", "tokenType"})
-    List<Token> findAllByAccountId(Integer id);
+    List<Token> findAllByAccountId(Long id);
 
     Optional<Token> findByAccessToken(String accessToken);
     Optional<Token> findByRefreshToken(String refreshToken);
     Optional<Token> findByUuidAccessTokenAndTokenScopeCategoryId(String uuidAccessToken, int tokenScopeCategoryId);
     Optional<Token> findByAccessTokenAndTokenScopeCategoryId(String accessToken, int tokenScopeCategoryId);
-    Optional<Token> findByAccountIdAndTokenScopeCategoryId(int accountId, int tokenScopeCategoryId);
+    Optional<Token> findByAccountIdAndTokenScopeCategoryId(Long accountId, int tokenScopeCategoryId);
 
 }
